backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==4.2.19
django-allauth==65.4.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
